refactor: dto 관련 클래스 좀 더 단순하게 정리
